Split out LED control + key handler
