+++ apps/main/key_handler.h
@@ -1,0 +1,7 @@
+#ifndef KEY_HANDLER_H_
+#define KEY_HANDLER_H_
+
+void app_key_init(void);
+
+void app_key_init_on_charging(void);
+#endif
