+++ apps/main/led_control.h
@@ -1,0 +1,9 @@
+#ifndef LED_CONTROL_H_
+#define LED_CONTROL_H_
+
+void LED_statusinit(void);
+void startLED_status(int ms);
+void stopLED_status(void);
+void LED_statusfun(const void *);
+
+#endif
